Make StopConsuming() more aggressive
Don't finish consuming unacked deliveries in deliveryChan.
Only finish deliveries on which we already called Consume().

This is to avoid the following issue: Consider the heartbeat dies so we
stop all consuming. Active Consume() calls are likely to run into errors
when trying to Ack(). We want to retry acking in those cases to not lose
the acks, which will in turn block the unacked deliveries (in
deliveryChan) from being passed to Consume(). If later the redis
connection recovers we might call Consume() on previously unacked
deliveries, even though those are likely to have been returned to ready
by the cleaner. So we should avoid conuming those.